Add Information and Generic Tiles
