+++ draft-packages/tile/KaizenDraft/Tile/MultiActionTile.scss
@@ -1,0 +1,6 @@
+@import "~@kaizen/design-tokens/sass/spacing";
+
+.actions {
+  padding-top: $kz-spacing-sm;
+  display: flex;
+}
